Feat: Add permission middleware to iam service

+++ services/iam-service/routes/api.php
@@ -14,17 +14,17 @@ Route::prefix('auth')->group(function () {
 });
 
 Route::prefix('users')->middleware('jwt.auth')->group(function () {
-    Route::post('/', [UserController::class, 'store']);
+    Route::post('/', [UserController::class, 'store'])->middleware('permission:user:invite');
     Route::get('/', [UserController::class, 'index']);
     Route::get('/{id}', [UserController::class, 'show']);
-    Route::put('/{id}', [UserController::class, 'update']);
-    Route::delete('/{id}', [UserController::class, 'destroy']);
+    Route::put('/{id}', [UserController::class, 'update'])->middleware('permission:user:update');
+    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('permission:user:delete');
 });
 
 Route::prefix('roles')->middleware('jwt.auth')->group(function () {
     Route::get('/', [RoleController::class, 'index']);
     Route::get('/user/{userId}', [RoleController::class, 'getUserRoles']);
-    Route::post('/assign', [RoleController::class, 'assignRole']);
-    Route::post('/remove', [RoleController::class, 'removeRole']);
+    Route::post('/assign', [RoleController::class, 'assignRole'])->middleware('permission:user:update');
+    Route::post('/remove', [RoleController::class, 'removeRole'])->middleware('permission:user:update');
 });
